move dep mount creation logic for testability
