+++ modal/mount.py
@@ -810,3 +810,110 @@ def _is_modal_path(remote_path: PurePosixPath):
         if is_modal_path:
             return True
     return False
+
+REMOTE_PACKAGES_PATH = "/__modal/deps"
+REMOTE_SITECUSTOMIZE_PATH = "/pkg/sitecustomize.py"
+
+SITECUSTOMIZE_CONTENT = f"""
+# This file is automatically generated by Modal.
+# It ensures that Modal's python dependencies are available in the Python PATH,
+# while prioritizing user-installed packages.
+import sys; sys.path.append('{REMOTE_PACKAGES_PATH}')
+""".strip()
+
+async def _create_single_mount(
+    client: _Client,
+    builder_version: str,
+    python_version: str,
+    platform: str,
+    arch: str,
+    uv_python_platform: str = None,
+    check_if_exists: bool = True,
+):
+    import subprocess
+    import tempfile
+
+    profile_environment = config.get("environment")
+
+    abi_tag = "cp" + python_version.replace(".", "")
+    mount_name = f"{builder_version}-{abi_tag}-{platform}-{arch}"
+    uv_python_platform = uv_python_platform or f"{arch}-{platform}"
+
+    if check_if_exists:
+        try:
+            Mount.from_name(mount_name, namespace=api_pb2.DEPLOYMENT_NAMESPACE_GLOBAL).hydrate(client)
+            print(f"✅ Found existing mount {mount_name} in global namespace.")
+            return
+        except modal.exception.NotFoundError:
+            pass
+
+    with tempfile.TemporaryDirectory() as tmpd:
+        print(f"📦 Building {mount_name}.")
+        requirements = os.path.join(os.path.dirname(__file__), f"requirements/{builder_version}.txt")
+        subprocess.run(
+            [
+                "uv",
+                "pip",
+                "install",
+                "--strict",
+                "--no-deps",
+                "--no-cache",
+                "-r",
+                requirements,
+                "--compile-bytecode",
+                "--target",
+                tmpd,
+                "--python-platform",
+                uv_python_platform,
+                "--python-version",
+                python_version,
+            ],
+            check=True,
+            capture_output=True,
+        )
+
+        print(f"🌐 Downloaded and unpacked packages to {tmpd}.")
+
+        python_mount = Mount._from_local_dir(tmpd, remote_path=REMOTE_PACKAGES_PATH)
+
+        with tempfile.NamedTemporaryFile() as sitecustomize:
+            sitecustomize.write(
+                SITECUSTOMIZE_CONTENT.encode("utf-8"),
+            )
+            sitecustomize.flush()
+
+            python_mount = python_mount.add_local_file(
+                sitecustomize.name,
+                remote_path=REMOTE_SITECUSTOMIZE_PATH,
+            )
+
+            await python_mount._deploy.aio(
+                mount_name,
+                api_pb2.DEPLOYMENT_NAMESPACE_GLOBAL,
+                environment_name=profile_environment,
+                client=client,
+            )
+            print(f"✅ Deployed mount {mount_name} to global namespace.")
+
+
+async def _create_client_dependency_mounts(client=None, check_if_exists=True):
+    for python_version in PYTHON_STANDALONE_VERSIONS:
+        # glibc >= 2.17
+        await _create_single_mount(
+            client,
+            "PREVIEW",
+            python_version,
+            "manylinux_2_17",
+            "x86_64",
+            check_if_exists=check_if_exists,
+        )
+        # musl >= 1.2
+        await _create_single_mount(
+            client,
+            "PREVIEW",
+            python_version,
+            "musllinux_1_2",
+            "x86_64",
+            uv_python_platform="x86_64-unknown-linux-musl",
+            check_if_exists=check_if_exists,
+        )
