+++ modal_global_objects/mounts/modal_client_dependencies.py
@@ -5,117 +5,15 @@ import tempfile
 
 from modal.client import Client
 from modal.config import config
-from modal.exception import NotFoundError
+
 from modal.image import ImageBuilderVersion, _get_modal_requirements_path
-from modal.mount import (
-    PYTHON_STANDALONE_VERSIONS,
-    Mount,
-)
+from modal.mount import _create_client_dependency_mounts
 from modal_proto import api_pb2
 
-REMOTE_PACKAGES_PATH = "/__modal/deps"
-REMOTE_SITECUSTOMIZE_PATH = "/pkg/sitecustomize.py"
-
-SITECUSTOMIZE_CONTENT = f"""
-# This file is automatically generated by Modal.
-# It ensures that Modal's python dependencies are available in the Python PATH,
-# while prioritizing user-installed packages.
-import sys; sys.path.append('{REMOTE_PACKAGES_PATH}')
-""".strip()
-
-async def _create_single_mount(
-    client: Client,
-    builder_version: ImageBuilderVersion,
-    python_version: str,
-    platform: str,
-    arch: str,
-    uv_python_platform: str = None,
-):
-    profile_environment = config.get("environment")
-
-    abi_tag = "cp" + python_version.replace(".", "")
-    mount_name = f"{builder_version}-{abi_tag}-{platform}-{arch}"
-    uv_python_platform = uv_python_platform or f"{arch}-{platform}"
-
-    try:
-        await Mount.from_name(mount_name, namespace=api_pb2.DEPLOYMENT_NAMESPACE_GLOBAL).hydrate(client)
-        print(f"✅ Found existing mount {mount_name} in global namespace.")
-        return
-    except NotFoundError:
-        pass
-
-    with tempfile.TemporaryDirectory() as tmpd:
-        print(f"📦 Building {mount_name}.")
-        requirements = _get_modal_requirements_path(builder_version)
-        subprocess.run(
-            [
-                "uv",
-                "pip",
-                "install",
-                "--strict",
-                "--no-deps",
-                "--no-cache",
-                "-r",
-                requirements,
-                "--compile-bytecode",
-                "--target",
-                tmpd,
-                "--python-platform",
-                uv_python_platform,
-                "--python-version",
-                python_version,
-            ],
-            check=True,
-            capture_output=True,
-        )
-
-        print(f"🌐 Downloaded and unpacked packages to {tmpd}.")
-
-        python_mount = Mount._from_local_dir(tmpd, remote_path=REMOTE_PACKAGES_PATH)
-
-        with tempfile.NamedTemporaryFile() as sitecustomize:
-            sitecustomize.write(
-                SITECUSTOMIZE_CONTENT.encode("utf-8"),
-            )
-            sitecustomize.flush()
-
-            python_mount = python_mount.add_local_file(
-                sitecustomize.name,
-                remote_path=REMOTE_SITECUSTOMIZE_PATH,
-            )
-
-            await python_mount._deploy.aio(
-                mount_name,
-                api_pb2.DEPLOYMENT_NAMESPACE_GLOBAL,
-                environment_name=profile_environment,
-                client=client,
-            )
-            print(f"✅ Deployed mount {mount_name} to global namespace.")
-
-
-async def create_client_dependency_mounts(client=None):
-    for python_version in PYTHON_STANDALONE_VERSIONS:
-        # glibc >= 2.17
-        await _create_single_mount(
-            client,
-            "PREVIEW",
-            python_version,
-            "manylinux_2_17",
-            "x86_64",
-        )
-        # musl >= 1.2
-        await _create_single_mount(
-            client,
-            "PREVIEW",
-            python_version,
-            "musllinux_1_2",
-            "x86_64",
-            uv_python_platform="x86_64-unknown-linux-musl"
-        )
 
 def main():
     loop = asyncio.get_event_loop()
-    loop.run_until_complete(create_client_dependency_mounts())
+    loop.run_until_complete(_create_client_dependency_mounts())
 
 if __name__ == "__main__":
     main()
